requirements.txt
numpy
pandas
matplotlib
scipy
scikit-learn
joblib
python-dotenv
torch
torchvision
torchaudio
stable-baselines3
gymnasium
tensorboard
yfinance
alpaca-trade-api
requests
ccxt
pymongo
sqlalchemy
psycopg2-binary
ephem
astroquery
ta
backtrader
vectorbt
flask
flask-cors
fastapi
uvicorn
dash
plotly
loguru
rich
